refactor(): Rename env package to getenv
